backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
python-dotenv==1.0.1
langchain-openai==0.0.8
openai==1.51.2
groq==0.37.1
httpx==0.27.2
edge-tts==7.2.7
azure-search-documents==11.5.2
azure-identity==1.17.1
langchain-community==0.0.28
langchain-text-splitters==0.0.1
pypdf==4.3.1
